refactor(jest): switch to `@swc/jest` from `ts-jest`
