+++ package.json
@@ -54,6 +54,7 @@
 		"@storybook/preact-webpack5": "^8.6.3",
 		"@storybook/react": "^8.6.3",
 		"@swc/core": "^1.11.5",
+		"@swc/jest": "^0.2.37",
 		"@swc/plugin-emotion": "^9.0.0",
 		"@testing-library/cypress": "^10.0.3",
 		"@testing-library/jest-dom": "^6.6.3",
@@ -104,7 +105,6 @@
 		"storybook": "^8.6.3",
 		"swc-loader": "^0.2.6",
 		"terser-webpack-plugin": "^5.3.12",
-		"ts-jest": "^29.2.6",
 		"ts-node": "^10.9.2",
 		"typescript": "^5.8.2",
 		"typescript-eslint": "^8.25.0",
