+++ src/client/__tests__/MainForm.test.tsx
@@ -1,9 +1,7 @@
 /**
  * @jest-environment jsdom
  */
-/** @jsx jsx */
 
-import { jsx } from '@emotion/react';
 import { act, fireEvent, render, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom/jest-globals';
 import { MainForm, MainFormProps } from '../components/MainForm';
